refactor(watchlist): 내부 클래스로 선언된 dto를 외부 클래스로 수정
