+++ src/main/java/com/sparta/spartatigers/domain/watchlist/dto/MatchDto.java
@@ -1,0 +1,13 @@
+package com.sparta.spartatigers.domain.watchlist.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class MatchDto {
+
+    private Long id;
+}
