+++ src/main/java/com/sparta/spartatigers/domain/watchlist/dto/request/CreateWatchListRequestDto.java
@@ -4,6 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import com.sparta.spartatigers.domain.watchlist.dto.MatchDto;
+import com.sparta.spartatigers.domain.watchlist.dto.RecordDto;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,21 +14,4 @@ public class CreateWatchListRequestDto {
 
     private MatchDto match;
     private RecordDto record;
-
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class MatchDto {
-
-        private Long id;
-    }
-
-    @Getter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class RecordDto {
-
-        private String content;
-        private int rate;
-    }
 }
