+++ src/main/java/com/sparta/spartatigers/domain/watchlist/dto/response/WatchListResponseDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import com.sparta.spartatigers.domain.match.dto.MatchScheduleDto;
-import com.sparta.spartatigers.domain.watchlist.dto.response.CreateWatchListResponseDto.RecordDto;
+import com.sparta.spartatigers.domain.watchlist.dto.RecordDto;
 import com.sparta.spartatigers.domain.watchlist.model.entity.WatchList;
 
 @Getter
@@ -14,11 +14,11 @@ import com.sparta.spartatigers.domain.watchlist.model.entity.WatchList;
 public class WatchListResponseDto {
     private Long id;
     private RecordDto record;
-    private MatchScheduleDto matchSchedule;
+    private MatchScheduleDto match;
 
     public static WatchListResponseDto of(WatchList watchList) {
         MatchScheduleDto matchScheduleDto = MatchScheduleDto.of(watchList.getMatch());
-        RecordDto recordDto = RecordDto.of(watchList);
-        return new WatchListResponseDto(watchList.getId(), recordDto, matchScheduleDto);
+        return new WatchListResponseDto(
+                watchList.getId(), RecordDto.of(watchList), matchScheduleDto);
     }
 }
